Improve php-cli handling of paths.conf, add covergage targets.
Case EA-5741:

This commit removes memory leaks in the handling of the paths.conf
file.  Also, it improves the ability of the parser to sensibly
recover from whitespace inclusion.  Finally, it places the parsing
of paths.conf under unit test.

This commit also adds code coverage targets to the build.  To
generate a coverage report, please refer to the README.apt file.