Add support for -ea_reference_dir option.
Case EA-7935: We need this for so that we can automatically
detect the correct PHP version for a given domain even when running
a PHP application that's not installed in the document root for
the domain.

The current use case for this is with wp-cli, which is installed
under /usr/local/cpanel/3rdparty and executed from cPanel (not
Apache) using php-cli.  We need to make sure it gets executed with
the correct version of PHP.

Also corrected previous date in changelog that the RPM build was
complaining about because Aug 9 2017 was not a Tuesday.

Team-Story: LC_9708
Team-Merge-Task: LC-9871