ZC-4400: rewrite to make it easier to work with and fix all the bugs
No matter how future versions of cpanel do the setting (.htaccess format,
move config + rendered Include template, etc) this will continue to work
on all version because its asking cPanel for the information. So as long
as the the lookup/caching continues to work this will do the right thing.