ZC-11159, ZC-11178: Eliminate $pkg in proc_args()
The proc_args() function returns whatever it believes to be PHP
version-package requested by the user. However, since EA-7961, the
function always returns the uninitialized variable. That value is now
always determined by the next function call. Since completely removing
the parameter in the returned list is going to change so much about the
test, which refers to return values by position in the list, just
eliminate the variable and hardcode an undef into the return in its
place.